+++ OwnTracks/OwnTracks/Subscriptions.h
@@ -1,0 +1,17 @@
+//
+//  Subscriptions.h
+//  OwnTracks
+//
+//  Created by Christoph Krey on 20.10.15.
+//  Copyright © 2015 OwnTracks. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <StoreKit/StoreKit.h>
+
+@interface Subscriptions : NSObject <SKProductsRequestDelegate, SKPaymentTransactionObserver, SKRequestDelegate>
++ (Subscriptions *)sharedInstance;
+@property (readonly, strong, nonatomic) NSNumber *recording;
+- (void)payRecording;
+
+@end